Refactored several lines of code, mainly the timer and some of GamePanel
The timer class was programmed with far too many lines of code and
accumulated a buildup of threads every time start or restart was called.
The new implementation is significantly skinnier and faces less problems
than the previous version. Overall, it is a much easier to read class
and no longer creates extra threads.